+++ Backend/ent/schema/review.go
@@ -20,7 +20,7 @@ func (Review) Fields() []ent.Field {
 	return []ent.Field{
 		field.UUID("id", uuid.UUID{}).
 			Default(uuid.New).Unique().StorageKey("id").Immutable(),
-		field.Float32("score").Range(0.0, 5.0).Nillable().Optional(),
+		field.Int8("score").Range(0.0, 5.0).Nillable().Optional(),
 		field.String("review_msg").Optional().Nillable(),
 		field.Time("review_time_at").Default(time.Now).Immutable(),
 	}
